Train code split, Part 2: moving some of the model.py code into clustering.py (#8)
Moving the clustering code into its own clustering.py module and adding docstrings.

Also, moving some common parameter type definitions to a params.py module.
