+++ src/midst_toolkit/models/clavaddpm/model.py
@@ -3,8 +3,8 @@ import json
 import math
 import pickle
 from abc import ABC, abstractmethod
-from collections import Counter, defaultdict
-from collections.abc import Callable, Generator, Iterator
+from collections import Counter
+from collections.abc import Callable, Generator
 from copy import deepcopy
 from dataclasses import astuple, dataclass, replace
 from pathlib import Path
@@ -18,7 +18,6 @@ import torch.nn.functional as F
 # ruff: noqa: N812
 from category_encoders import LeaveOneOutEncoder
 from scipy.special import expit, softmax
-from sklearn.cluster import KMeans
 from sklearn.impute import SimpleImputer
 from sklearn.metrics import (
     classification_report,
@@ -26,7 +25,6 @@ from sklearn.metrics import (
     r2_score,
     roc_auc_score,
 )
-from sklearn.mixture import BayesianGaussianMixture, GaussianMixture
 from sklearn.model_selection import train_test_split
 from sklearn.pipeline import make_pipeline
 from sklearn.preprocessing import (
@@ -269,377 +267,6 @@ class Classifier(nn.Module):
         return self.model(x)
 
 
-def pair_clustering_keep_id(
-    child_df: pd.DataFrame,
-    child_domain_dict: dict[str, Any],
-    parent_df: pd.DataFrame,
-    parent_domain_dict: dict[str, Any],
-    child_primary_key: str,
-    parent_primary_key: str,
-    num_clusters: int,
-    parent_scale: float,
-    key_scale: float,
-    parent_name: str,
-    child_name: str,
-    clustering_method: str = "kmeans",
-) -> tuple[pd.DataFrame, pd.DataFrame, dict[int, dict[int, float]]]:
-    original_child_cols = list(child_df.columns)
-    original_parent_cols = list(parent_df.columns)
-
-    relation_cluster_name = f"{parent_name}_{child_name}_cluster"
-
-    child_data = child_df.to_numpy()
-    parent_data = parent_df.to_numpy()
-
-    child_num_cols = []
-    child_cat_cols = []
-
-    parent_num_cols = []
-    parent_cat_cols = []
-
-    for col_index, col in enumerate(original_child_cols):
-        if col in child_domain_dict:
-            if child_domain_dict[col]["type"] == "discrete":
-                child_cat_cols.append((col_index, col))
-            else:
-                child_num_cols.append((col_index, col))
-
-    for col_index, col in enumerate(original_parent_cols):
-        if col in parent_domain_dict:
-            if parent_domain_dict[col]["type"] == "discrete":
-                parent_cat_cols.append((col_index, col))
-            else:
-                parent_num_cols.append((col_index, col))
-
-    parent_primary_key_index = original_parent_cols.index(parent_primary_key)
-    foreing_key_index = original_child_cols.index(parent_primary_key)
-
-    # sort child data by foreign key
-    sorted_child_data = child_data[np.argsort(child_data[:, foreing_key_index])]
-    child_group_data_dict = get_group_data_dict(
-        sorted_child_data,
-        [
-            foreing_key_index,
-        ],
-    )
-
-    # sort parent data by primary key
-    sorted_parent_data = parent_data[np.argsort(parent_data[:, parent_primary_key_index])]
-
-    group_lengths = []
-    unique_group_ids = sorted_parent_data[:, parent_primary_key_index]
-    for group_id in unique_group_ids:
-        group_id = tuple([group_id])
-        # ruff: noqa: C409
-        if group_id not in child_group_data_dict:
-            group_lengths.append(0)
-        else:
-            group_lengths.append(len(child_group_data_dict[group_id]))
-
-    group_lengths_np = np.array(group_lengths, dtype=int)
-
-    sorted_parent_data_repeated = np.repeat(sorted_parent_data, group_lengths_np, axis=0)
-    assert (sorted_parent_data_repeated[:, parent_primary_key_index] == sorted_child_data[:, foreing_key_index]).all()
-
-    child_group_data = get_group_data(
-        sorted_child_data,
-        [
-            foreing_key_index,
-        ],
-    )
-
-    sorted_child_num_data = sorted_child_data[:, [col_index for col_index, col in child_num_cols]]
-    sorted_child_cat_data = sorted_child_data[:, [col_index for col_index, col in child_cat_cols]]
-    sorted_parent_num_data = sorted_parent_data_repeated[:, [col_index for col_index, col in parent_num_cols]]
-    sorted_parent_cat_data = sorted_parent_data_repeated[:, [col_index for col_index, col in parent_cat_cols]]
-
-    joint_num_matrix = np.concatenate([sorted_child_num_data, sorted_parent_num_data], axis=1)
-    joint_cat_matrix = np.concatenate([sorted_child_cat_data, sorted_parent_cat_data], axis=1)
-
-    if joint_cat_matrix.shape[1] > 0:
-        joint_cat_matrix_p_index = sorted_child_cat_data.shape[1]
-        joint_num_matrix_p_index = sorted_child_num_data.shape[1]
-
-        cat_converted = []
-        label_encoders = []
-        for i in range(joint_cat_matrix.shape[1]):
-            # A threshold of 1000 unique values is used to prevent the one-hot encoding of large categorical columns
-            if len(np.unique(joint_cat_matrix[:, i])) > 1000:
-                continue
-            label_encoder = LabelEncoder()
-            cat_converted.append(label_encoder.fit_transform(joint_cat_matrix[:, i]).astype(float))
-            label_encoders.append(label_encoder)
-
-        cat_converted_transposed = np.vstack(cat_converted).T
-
-        # Initialize an empty array to store the encoded values
-        cat_one_hot = np.empty((cat_converted_transposed.shape[0], 0))
-
-        # Loop through each column in the data and encode it
-        for col in range(cat_converted_transposed.shape[1]):
-            encoder = OneHotEncoder(sparse_output=False)
-            column = cat_converted_transposed[:, col].reshape(-1, 1)
-            encoded_column = encoder.fit_transform(column)
-            cat_one_hot = np.concatenate((cat_one_hot, encoded_column), axis=1)
-
-        cat_one_hot[:, joint_cat_matrix_p_index:] = parent_scale * cat_one_hot[:, joint_cat_matrix_p_index:]
-
-    # Perform quantile normalization using QuantileTransformer
-    num_quantile = quantile_normalize_sklearn(joint_num_matrix)
-    num_min_max = min_max_normalize_sklearn(joint_num_matrix)
-
-    # key_quantile =
-    #   quantile_normalize_sklearn(sorted_parent_data_repeated[:, parent_primary_key_index].reshape(-1, 1))
-    key_min_max = min_max_normalize_sklearn(sorted_parent_data_repeated[:, parent_primary_key_index].reshape(-1, 1))
-
-    # key_scaled = key_scaler * key_quantile
-    key_scaled = key_scale * key_min_max
-
-    num_quantile[:, joint_num_matrix_p_index:] = parent_scale * num_quantile[:, joint_num_matrix_p_index:]
-    num_min_max[:, joint_num_matrix_p_index:] = parent_scale * num_min_max[:, joint_num_matrix_p_index:]
-
-    if joint_cat_matrix.shape[1] > 0:
-        cluster_data = np.concatenate((num_min_max, cat_one_hot, key_scaled), axis=1)
-    else:
-        cluster_data = np.concatenate((num_min_max, key_scaled), axis=1)
-
-    child_group_lengths = np.array([len(group) for group in child_group_data], dtype=int)
-    num_clusters = min(num_clusters, len(cluster_data))
-
-    # print('clustering')
-    if clustering_method == "kmeans":
-        kmeans = KMeans(n_clusters=num_clusters, n_init="auto", init="k-means++")
-        kmeans.fit(cluster_data)
-        cluster_labels = kmeans.labels_
-    elif clustering_method == "both":
-        gmm = GaussianMixture(
-            n_components=num_clusters,
-            verbose=1,
-            covariance_type="diag",
-            init_params="k-means++",
-            tol=0.0001,
-        )
-        gmm.fit(cluster_data)
-        cluster_labels = gmm.predict(cluster_data)
-    elif clustering_method == "variational":
-        gmm = BayesianGaussianMixture(
-            n_components=num_clusters,
-            verbose=1,
-            covariance_type="diag",
-            init_params="k-means++",
-            tol=0.0001,
-        )
-        gmm.fit(cluster_data)
-        cluster_labels = gmm.predict_proba(cluster_data)
-    elif clustering_method == "gmm":
-        gmm = GaussianMixture(
-            n_components=num_clusters,
-            verbose=1,
-            covariance_type="diag",
-        )
-        gmm.fit(cluster_data)
-        cluster_labels = gmm.predict(cluster_data)
-
-    if clustering_method == "variational":
-        group_cluster_labels, agree_rates = aggregate_and_sample(cluster_labels, child_group_lengths)
-    else:
-        # voting to determine the cluster label for each parent
-        group_cluster_labels = []
-        curr_index = 0
-        agree_rates = []
-        for group_length in child_group_lengths:
-            # First, determine the most common label in the current group
-            most_common_label_count = np.max(np.bincount(cluster_labels[curr_index : curr_index + group_length]))
-            group_cluster_label = np.argmax(np.bincount(cluster_labels[curr_index : curr_index + group_length]))
-            group_cluster_labels.append(int(group_cluster_label))
-
-            # Compute agree rate using the most common label count
-            agree_rate = most_common_label_count / group_length
-            agree_rates.append(agree_rate)
-
-            # Then, update the curr_index for the next iteration
-            curr_index += group_length
-
-    # Compute the average agree rate across all groups
-    average_agree_rate = np.mean(agree_rates)
-    print("Average agree rate: ", average_agree_rate)
-
-    group_assignment = np.repeat(group_cluster_labels, child_group_lengths, axis=0).reshape((-1, 1))
-
-    # obtain the child data with clustering
-    sorted_child_data_with_cluster = np.concatenate([sorted_child_data, group_assignment], axis=1)
-
-    group_labels_list = group_cluster_labels
-    group_lengths_list = child_group_lengths.tolist()
-
-    group_lengths_dict: dict[int, dict[int, int]] = {}
-    for i in range(len(group_labels_list)):
-        group_label = group_labels_list[i]
-        if group_label not in group_lengths_dict:
-            group_lengths_dict[group_label] = defaultdict(int)
-        group_lengths_dict[group_label][group_lengths_list[i]] += 1
-
-    group_lengths_prob_dicts: dict[int, dict[int, float]] = {}
-    for group_label, freq_dict in group_lengths_dict.items():
-        group_lengths_prob_dicts[group_label] = freq_to_prob(freq_dict)
-
-    # recover the preprocessed data back to dataframe
-    child_df_with_cluster = pd.DataFrame(
-        sorted_child_data_with_cluster,
-        columns=original_child_cols + [relation_cluster_name],
-    )
-
-    # recover child df order
-    child_df_with_cluster = pd.merge(
-        child_df[[child_primary_key]],
-        child_df_with_cluster,
-        on=child_primary_key,
-        how="left",
-    )
-
-    parent_id_to_cluster: dict[Any, Any] = {}
-    for i in range(len(sorted_child_data)):
-        parent_id = sorted_child_data[i, foreing_key_index]
-        if parent_id in parent_id_to_cluster:
-            assert parent_id_to_cluster[parent_id] == sorted_child_data_with_cluster[i, -1]
-            continue
-        parent_id_to_cluster[parent_id] = sorted_child_data_with_cluster[i, -1]
-
-    max_cluster_label = max(parent_id_to_cluster.values())
-
-    parent_data_clusters = []
-    for i in range(len(parent_data)):
-        if parent_data[i, parent_primary_key_index] in parent_id_to_cluster:
-            parent_data_clusters.append(parent_id_to_cluster[parent_data[i, parent_primary_key_index]])
-        else:
-            parent_data_clusters.append(max_cluster_label + 1)
-
-    parent_data_clusters_np = np.array(parent_data_clusters).reshape(-1, 1)
-    parent_data_with_cluster = np.concatenate([parent_data, parent_data_clusters_np], axis=1)
-    parent_df_with_cluster = pd.DataFrame(
-        parent_data_with_cluster, columns=original_parent_cols + [relation_cluster_name]
-    )
-
-    new_col_entry = {
-        "type": "discrete",
-        "size": len(set(parent_data_clusters_np.flatten())),
-    }
-
-    print("Number of cluster centers: ", len(set(parent_data_clusters_np.flatten())))
-
-    parent_domain_dict[relation_cluster_name] = new_col_entry.copy()
-    child_domain_dict[relation_cluster_name] = new_col_entry.copy()
-
-    return parent_df_with_cluster, child_df_with_cluster, group_lengths_prob_dicts
-
-
-def get_group_data_dict(
-    np_data: np.ndarray,
-    group_id_attrs: list[int] | None = None,
-) -> dict[tuple[Any, ...], list[np.ndarray]]:
-    if group_id_attrs is None:
-        group_id_attrs = [0]
-
-    group_data_dict: dict[tuple[Any, ...], list[np.ndarray]] = {}
-    data_len = len(np_data)
-    for i in range(data_len):
-        row_id = tuple(np_data[i, group_id_attrs])
-        if row_id not in group_data_dict:
-            group_data_dict[row_id] = []
-        group_data_dict[row_id].append(np_data[i])
-
-    return group_data_dict
-
-
-def get_group_data(
-    np_data: np.ndarray,
-    group_id_attrs: list[int] | None = None,
-) -> np.ndarray:
-    if group_id_attrs is None:
-        group_id_attrs = [0]
-
-    group_data_list = []
-    data_len = len(np_data)
-    i = 0
-    while i < data_len:
-        group = []
-        row_id = np_data[i, group_id_attrs]
-
-        while (np_data[i, group_id_attrs] == row_id).all():
-            group.append(np_data[i])
-            i += 1
-            if i >= data_len:
-                break
-        group_data_list.append(np.array(group))
-
-    return np.array(group_data_list, dtype=object)
-
-
-def quantile_normalize_sklearn(matrix: np.ndarray) -> np.ndarray:
-    transformer = QuantileTransformer(
-        output_distribution="normal", random_state=42
-    )  # Change output_distribution as needed
-
-    normalized_data = np.empty((matrix.shape[0], 0))
-
-    # Apply QuantileTransformer to each column and concatenate the results
-    for col in range(matrix.shape[1]):
-        column = matrix[:, col].reshape(-1, 1)
-        transformed_column = transformer.fit_transform(column)
-        normalized_data = np.concatenate((normalized_data, transformed_column), axis=1)
-
-    return normalized_data
-
-
-def min_max_normalize_sklearn(matrix: np.ndarray) -> np.ndarray:
-    scaler = MinMaxScaler(feature_range=(-1, 1))
-
-    normalized_data = np.empty((matrix.shape[0], 0))
-
-    # Apply MinMaxScaler to each column and concatenate the results
-    for col in range(matrix.shape[1]):
-        column = matrix[:, col].reshape(-1, 1)
-        transformed_column = scaler.fit_transform(column)
-        normalized_data = np.concatenate((normalized_data, transformed_column), axis=1)
-
-    return normalized_data
-
-
-def aggregate_and_sample(
-    cluster_probabilities: np.ndarray,
-    child_group_lengths: np.ndarray,
-) -> tuple[list[int], list[float]]:
-    group_cluster_labels = []
-    curr_index = 0
-    agree_rates = []
-
-    for group_length in child_group_lengths:
-        # Aggregate the probability distributions by taking the mean
-        group_probability_distribution = np.mean(cluster_probabilities[curr_index : curr_index + group_length], axis=0)
-
-        # Sample the label from the aggregated distribution
-        group_cluster_label = np.random.choice(
-            range(len(group_probability_distribution)), p=group_probability_distribution
-        )
-        group_cluster_labels.append(group_cluster_label)
-
-        # Compute the max probability as the agree rate
-        max_probability = np.max(group_probability_distribution)
-        agree_rates.append(max_probability)
-
-        # Update the curr_index for the next iteration
-        curr_index += group_length
-
-    return group_cluster_labels, agree_rates
-
-
-def freq_to_prob(freq_dict: dict[int, int]) -> dict[int, float]:
-    prob_dict: dict[Any, float] = {}
-    for key, freq in freq_dict.items():
-        prob_dict[key] = freq / sum(list(freq_dict.values()))
-    return prob_dict
-
-
 def get_table_info(df: pd.DataFrame, domain_dict: dict[str, Any], y_col: str) -> dict[str, Any]:
     cat_cols = []
     num_cols = []
@@ -864,22 +491,6 @@ def get_model(
         return ResNetDiffusion(**model_params)
 
     raise ValueError("Unknown model!")
-
-
-def update_ema(
-    target_params: Iterator[nn.Parameter],
-    source_params: Iterator[nn.Parameter],
-    rate: float = 0.999,
-) -> None:
-    """
-    Update target parameters to be closer to those of source parameters using
-    an exponential moving average.
-    :param target_params: the target parameter sequence.
-    :param source_params: the source parameter sequence.
-    :param rate: the EMA rate (closer to 1 means slower).
-    """
-    for targ, src in zip(target_params, source_params):
-        targ.detach().mul_(rate).add_(src.detach(), alpha=1 - rate)
 
 
 class ScheduleSampler(ABC):
