+++ src/midst_toolkit/models/clavaddpm/typing.py
@@ -1,0 +1,8 @@
+from typing import Any
+
+
+# TODO: Temporary, will switch to classes later
+Configs = dict[str, Any]
+Tables = dict[str, dict[str, Any]]
+RelationOrder = list[tuple[str, str]]
+GroupLengthsProbDicts = dict[tuple[str, str], dict[int, dict[int, float]]]
